now can mint individual hostname certs.  sigh archive.org servers cant run hirsute+ due to oldness and networking - clamp to focal.  need to use --net=host now for archive.org issueing server 🤦‍♀️

+++ create-https-certs.sh
@@ -8,15 +8,27 @@
 # https://community.letsencrypt.org/t/confusing-on-root-domain-with-wildcard-cert/56113
 
 
-[ $# -lt 1 ]  &&  echo "usage: $0 [TLS_DOMAIN eg: x.archive.org]"  &&  exit 1
+[ $# -lt 1 ]  &&  echo "usage: $0 [TLS_DOMAIN eg: x.archive.org]\nusage: $0 -d [single hostname cert] -d [single hostname cert] ..."
+[ $# -lt 1 ]  &&  exit 1
 set -x
-TLS_DOMAIN=$1
+
+TLS_DOMAIN=
+SINGLETONS=
+[ $# -eq 1 ]  &&  TLS_DOMAIN=$1
+[ $# -gt 2 ]  &&  SINGLETONS=$@
+
+BASENAME=$TLS_DOMAIN
+[ $SINGLETONS ]  &&  BASENAME=$2
+
+[ $TLS_DOMAIN ]  &&  ARGS="--preferred-challenges  dns-01 -d '*.${TLS_DOMAIN?}'"
+[ $SINGLETONS ]  &&  ARGS="--preferred-challenges http-01       ${SINGLETONS?}"
+
 
 
 # This part is pretty slow, so let's do all the slow setup and save it for reuse.
 # You can `sudo docker rmi certomatic` later as desired
 sudo docker run -it --rm certomatic echo  || (
-  sudo docker run -it --name certomatic ubuntu:rolling bash -c '
+  sudo docker run -it --name certomatic --net=host ubuntu:focal bash -c '
     export DEBIAN_FRONTEND=noninteractive
     apt-get -yqq update
     apt-get -yqq install  certbot
@@ -26,18 +38,18 @@ sudo docker run -it --rm certomatic echo  || (
 )
 
 
-sudo touch     ${TLS_DOMAIN?}-{cert,key}.pem
-sudo chmod 666 ${TLS_DOMAIN?}-{cert,key}.pem
+sudo touch     ${BASENAME?}-{cert,key}.pem || echo 'nice.  moving on..'
+sudo chmod 666 ${BASENAME?}-{cert,key}.pem || echo 'nice.  moving on..'
 
-sudo docker run -it --rm -v $(pwd):/x  certomatic  bash -c "
-  certbot certonly --manual --manual-public-ip-logging-ok --preferred-challenges dns-01 \
-    --server https://acme-v02.api.letsencrypt.org/directory -d '*.${TLS_DOMAIN?}'
+sudo docker run -it --rm --net=host -v $(pwd):/x  certomatic  bash -c "
+  certbot certonly --manual --agree-tos --manual-public-ip-logging-ok  \
+    --server https://acme-v02.api.letsencrypt.org/directory $ARGS
   set -x
   cd /etc/letsencrypt/live/*/
-  cp -p fullchain.pem /x/${TLS_DOMAIN?}-cert.pem
-  cp -p privkey.pem   /x/${TLS_DOMAIN?}-key.pem
+  cp -p fullchain.pem /x/${BASENAME?}-cert.pem
+  cp -p privkey.pem   /x/${BASENAME?}-key.pem
   echo 'type exit to finish'
   bash
 "
-sudo chmod 444 ${TLS_DOMAIN?}-cert.pem
-sudo chmod 400 ${TLS_DOMAIN?}-key.pem
+sudo chmod 444 ${BASENAME?}-cert.pem
+sudo chmod 400 ${BASENAME?}-key.pem
